Extract from SCU and store FW version at boot phase
BZ: 112309

The goal is to store the complete FW version early during boot
via a SCU IPC request, so that the data can be read under
/sys/kernel/fw_update/fw_info/fw_version and properties set
accordingly in the .rc files.

When starting, this driver will send the IPCMSG_FW_REVISION message
and a buffer to SCU in order to get the installed firmware version
(IFWI, SCU, Chaabi...).
SCU fills the passed buffer. Content of the buffer is then parsed
and stored in specific ifwi, ia32fw, chaabi... versions buffers.

Those buffers are then returned to setprop_from_sysfs requests in
the .rc files.

Change-Id: I3afe00fa0bf8f35f79bd33ef6c213a405ddf7a70
Signed-off-by: Michel Jauffres <michelx.jauffres@intel.com>