[WORKAROUND] dwc3-device: Wait 300ms before resuming the controller
In Droidboot, sys.usb.config is initially set as charging mode, then
NONE, and finally adb mode. When it's in charging mode, DUT is connected
to host, and enumeration occurs. For USB2, bus enters L2 first and
hibernation takes place. Then host will reset the bus, but before this
happens, it's found that sys.usb.config is set NONE during which the
controller is resumed by SW. Seems there's a conflict between the SW
resume and bus resume, and PHY hangs.

The patch forces it to wait 300ms before resuming the controller to give
enough time for the bus to be resumed by host to avoid hitting the
corner case.

Change-Id: I186b5b4890712989e999038f83a00df262eba9bc
Tracked-On: https://jira01.devtools.intel.com/browse/AW-3045
Signed-off-by: jzhuan5 <jin.can.zhuang@intel.com>
Signed-off-by: Simon Dubray <simonx.dubray@intel.com>
Reviewed-on: https://android.intel.com/554963
Reviewed-by: Bel Aj Ali, HabibX <habibx.bel.aj.ali@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>