i2c: disable pull_up in i2c-6 for PVT
This patch is copyed from the following patch(the commit ID listed
below), but not 100% of it, the missing changes are:
1. delete file "arch/x86/platform/intel-mid/device_libs/platform_lsm6ds3.c"
2. delete file "drivers/iio/imu/st_lsm6ds3/st_lsm6ds3_i2c.c"
3. delete file "include/linux/platform_data/st_lsm6ds3_pdata.h"

commit ID for reference: 62306d317ad1e2a9890d260e087cbada2198e9e4

Note: They have to be compiled & get in together.

The pull_up is disabled before the first i2c access during probe() of
the 2 drivers using i2c-6.
Since we don't know which driver will be probed first, the pull_up must
be disabled in both drivers.

Change-Id: Ibc41bc2a7601002937dc647adc6b5974b56d66b6
Tracked-On: https://jira.ndg.intel.com/browse/MARVIN-795
Signed-off-by: Guillaume Marcenac <guillaumex.marcenac@intel.com>
Signed-off-by: Kai Qiang <kaix.qiang@intel.com>
Reviewed-on: https://android.intel.com/414249
Reviewed-by: Korpershoek, MattijsX <mattijsx.korpershoek@intel.com>
Reviewed-by: Akue, LoicX <loicx.akue@intel.com>
Reviewed-by: Masson, JulienX <julienx.masson@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>