display: re-enable reset after boot complete
We have removed the reset function to avoid flickering during boot.
But due to this change, the user space does not have the possibility to reset
the panel.

To resolve this issue the reset function is disabled by default, and is
enabled again at the end of dsi_controller_init().

Change-Id: Ib02578055263d719694791b7650bc4033d9a9710
Tracked-On: https://jira.ndg.intel.com/browse/MARVIN-1266
Signed-off-by: Guillaume Marcenac <guillaumex.marcenac@intel.com>
Reviewed-on: https://android.intel.com/456569
Reviewed-by: Rouis, KhalifaX <khalifax.rouis@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>