display: debugfs: propagate error code to userland
In the case of a debugfs_write() failure we are returning 0 to the
userspace. By returning 0, userspace knows that we were unable to
write but keeps retrying. It has no clue of the type of error which
happens (e.g. ENOMEM, EINVAL, ...)

This patch reworks debugfs_write() to transfer the error codes to the
userland.

Change-Id: I122a0a826c35e88e5e4e3b328682ace2ceb05a06
Tracked-On: https://jira.ndg.intel.com/browse/MARVIN-1315
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>
Reviewed-on: https://android.intel.com/459420
Reviewed-by: Masson, JulienX <julienx.masson@intel.com>
Reviewed-by: Mahe, Erwan <erwan.mahe@intel.com>
Reviewed-by: Hadjimegrian, MathieuX <mathieux.hadjimegrian@intel.com>
Reviewed-by: Gong, Sophia <sophia.gong@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>