iio: imu: st_lsm6ds3: disable irq in suspend
As required by the linux power management API, the IO operations must be
quiesced while suspended. This patch prevents irq from firing while
suspending to RAM (S3), and making access to the below I2C controller
which might already be in suspend.

There is no impact on the wake-up from S3 capability : the notion of
disabling an irq is orthogonal to the one of disabling a wake-up source,
and 2 different IPs address these 2 notions anyway :
 - wake-up controller for enable_irq_wake()
 - interrupt controller for enable_irq()

Change-Id: Ic077d34a49efb6ecda611abe8e1695fbe4ecf1e6
Tracked-On: https://jira.ndg.intel.com/browse/MARVIN-1413
Signed-off-by: Robert Jarzmik <robert.jarzmik@intel.com>
Reviewed-on: https://android.intel.com/474183
Reviewed-by: Ranquet, GuillaumeX <guillaumex.ranquet@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Yan, Like <like.yan@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>