charger: force the bq24232 charger type to USB
The healthd ignores power_supply entries with unknown type.
With this patch the charger's type becomes USB by default
if no cable is plugged in. This is done right after its registration
to the power_supply class in order to do not use the system USB
management and keep relying on its own charging management

Change-Id: Ic584ae4c9ae4cc773875769cdca4c20937c973ee
Signed-off-by: Antonio Casolaro <antoniox.casolaro@intel.com>
Signed-off-by: Guillaume Marcenac <guillaumex.marcenac@intel.com>
Tracked-On: https://jira.ndg.intel.com/browse/MARVIN-1378
Reviewed-on: https://android.intel.com/466947
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Ranquet, GuillaumeX <guillaumex.ranquet@intel.com>