Audio: Enable fake suspend for bt-a2dp codec-codec link
We want to enable S3 for the OS, while BT A2DP loopback UC is running.
This loopback doesn't involve IA as it is a codec-codec link.
In S3, a lot of userspace timers are suspended, and this results
in better power savings when compared to S0ix.

Add a new control to indicate to the driver that a BT A2DP UC is
going on. In this case the driver indicates that it doesn't want to
suspend the device using the pci_save_state() call. This means that
the S3 legacy suspend call returns true, indicating that the driver
is suspended, but the LPE is not powergated by SCU. From the SCU
point-of-view, it is still an LPMP3 flow, where the LPE is not
powergated. Hence the LPMP3 count gets incremented, not the S3 count.

Change-Id: I7c0ae20196d05f93d32e1ed91a20dfef90a3559d
Orig-Change-Id: I6a214b407ca9255a0e95db3ccb4ebbfdbad12b51
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-48508
Signed-off-by: Omair M Abdullah <omair.m.abdullah@intel.com>
Signed-off-by: Jyoti, Anand B <anand.b.jyoti@intel.com>
Signed-off-by: Praveen Diwakar <praveen.diwakar@intel.com>
Signed-off-by: dnyaneshwarx.r.jadhav@intel.com <dnyaneshwarx.r.jadhav@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/IAWP-69
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>
Reviewed-on: https://android.intel.com/476929
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>