ASoC: Intel: compensate bit slip on SSP0 flag
On CDP boards we have a clock mismatch between the DMIC and the on
SoC's DSP.
The DMIC expects a clock of 16kHz but the DSP provides 16,667kHz.
This is a SoC limitation, which can only derive of 19,2MHz for devices
plugged on SSP{0,1,2}.

Since the DMIC *must* have 64 bit clocks in each stereo frame, we have
the following clocking configuration:
- base clock: 19.2MHz
- bit clock cycles per frame: 64
- frame sync frequency: {8,16,48} kHz
  -> 19,2 MHz / (SCR*64) where SCR must be an integer and the result
  must be close to 8,16,48} kHz. We have chosen SCR=17 so:
  FS = 19,2MHz / ((17+1) * 64) ~= 16,666666666666668 kHz.

Since we should sample at 16KHz, we have clock drifting.
This is compensated within the DSP firmware.
To trigger the sample decimation, we have to request a
frame_size of 64 to the DSP when in i2s mode.

Change-Id: Ib20a1fbf0d23c7bea8164cae1bdee12199e69325
Tracked-On: https://jira01.devtools.intel.com/browse/IAWP-69
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>
Reviewed-on: https://android.intel.com/475216
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>