UPSTREAM: usbnet: cleanup after bind() in probe()
Bug: 28744625

In case bind() works, but a later error forces bailing
in probe() in error cases work and a timer may be scheduled.
They must be killed. This fixes an error case related to
the double free reported in
http://www.spinics.net/lists/netdev/msg367669.html
and needs to go on top of Linus' fix to cdc-ncm.

(cherry picked from commit 1666984c8625b3db19a9abc298931d35ab7bc64b)

Bug: 31554728
Change-Id: Iec7d659d903eab0adee1c91cb8f5b84e9775c09d
Tracked-On: https://jira01.devtools.intel.com/browse/AW-1795
Signed-off-by: Oliver Neukum <ONeukum@suse.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Yueyao (Nathan) Zhu <yueyao@google.com>
Signed-off-by: Victor Tasayco Loarte <victorx.tasayco.loarte@intel.com>
Reviewed-on: https://android.intel.com/541433
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>