als: force probe failure when ALS not available
In the current implementation, when the i2c device detected isn't
the one expected, the taos_probe() function still returns 0.
The runtime power management then tries to call the suspend and
resume callbacks even if the pm_runtime wasn't properly configured,
leading to a NULL pointer dereference.

Let's return -ENODEV to avoid this.

Change-Id: I9bc30917c58e06a76fa85e963a411fb585150ab6
Tracked-On: https://jira.ndg.intel.com/browse/MARVIN-1578
Signed-off-by: Loic Akue <loicx.akue@intel.com>
Reviewed-on: https://android.intel.com/482331
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Ranquet, GuillaumeX <guillaumex.ranquet@intel.com>
Reviewed-by: Bel Aj Ali, HabibX <habibx.bel.aj.ali@intel.com>
Reviewed-by: Yan, Like <like.yan@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>