em: bq25898 initial charger driver
Initial version of BQ25898 charger implementation (done for POC and bring
up purpose).
This driver needs to have the proper data about the bq25898 in the IFWI
to probe properly.
This driver contains a debug fs implementation for debug and dump information related
to the bq25898 charger.
Termination charge timer is embedded in the driver and will continue to charge
30mn after charge termination is done to leave enough time to HW team to take measurement.
The proper timing will be done under another patch.
There is also the battery monitor placeholder implementation included which will
be refined in an upcoming patch.

Change-Id: I6f624cd677096f2e14ab70e6b191e3b1be4d627d
Tracked-On: https://jira01.devtools.intel.com/browse/AW-17
Signed-off-by: kslimani <kamel.slimani@intel.com>
Signed-off-by: Marc Blassin <marc.blassin@intel.com>
Reviewed-on: https://android.intel.com/484466
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>