spid: add CDP_ALL macro
As we have no way to differentiate Phone/Tablets and CloverDale Peak
products, we end up putting all the CDP SPIDs with the INTEL_MID_BOARD
macro which is both painful and error prone.

Add a CloverDale Peak macro to ease the selection of all CDP based
products and platforms.

Change-Id: Ie99d55afc8b2af494649182c37d45c5df29bff30
Tracked-On: https://jira01.devtools.intel.com/browse/AW-195
Signed-off-by: Guillaume Ranquet <guillaume.ranquet@intel.com>
Reviewed-on: https://android.intel.com/491282
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Korpershoek, MattijsX <mattijsx.korpershoek@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>