soc_mrfld: Force lss-8 in D0i3
As we are diverging from the original Merrifield design, we are re-using
GPIO173 to enable the display LDO on shasta and anthracite product.
The problem is that GPIO173 is needed for HSIC.
If HSIC is unable to probe (because GPIO173 is missing or already
requested), the pci device associated will stay in D0 as there's no
drivers to put it in suspend... thus breaking S0i3.

Since we are not using HSIC on CDP, we are removing the entry in the
pci tables and thus force the lss associated to D0i3

Change-Id: I9a25b7874716bb1bad736c520ea40b0b56960a86
Tracked-On: https://jira01.devtools.intel.com/browse/AW-194
Signed-off-by: Guillaume Ranquet <guillaume.ranquet@intel.com>
Reviewed-on: https://android.intel.com/491283
Reviewed-by: Korpershoek, MattijsX <mattijsx.korpershoek@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>