power: pmic_ccsm: add a notifier
Add a notifier to signal battery status changes, overheat and temperature
zone. This will make the platform_data callback deprecated, and remove
the inter-driver direct calls when the notifier will be properly used.

The notifier is of atomic kind, ie. it is expected that the notification
will come in an interrupt path

Change-Id: Ia370e2c39cf2504d9ed8df7e7fab7df311409ff1
Tracked-On: https://jira01.devtools.intel.com/browse/AW-280
Signed-off-by: Robert Jarzmik <robert.jarzmik@intel.com>
Reviewed-on: https://android.intel.com/493921
Reviewed-by: Ranquet, Guillaume <guillaume.ranquet@intel.com>
Reviewed-by: Blassin, Marc <marc.blassin@intel.com>
Tested-by: Blassin, Marc <marc.blassin@intel.com>
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>