em: kernel panic when enabling bq25898 debug
PM runtime functions are called very early in the process.
When enabling debug, dev_dbg output is using variables which may
not have been yet initialised, thus would cause a kernel panic.
This patch prevent dev_dbg to be called in case of NULL pointer.
The second issue fixed by this patch is the order of the i2c_set_clientdata
and the pm_runtime_enable. PM runtime uses the clientdata and if this is not
set before calling pm_runtime_enable, we may have calls to pm using unitialized
data (which anyway will be harmless after this patch applied.

Change-Id: I5d53fcdc7aee35860862d577a739d6adba225b0b
Tracked-On: https://jira01.devtools.intel.com/browse/AW-364
Signed-off-by: Marc Blassin <marc.blassin@intel.com>
Reviewed-on: https://android.intel.com/498062
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>