gps: disable host_wake ISR in intel_mid_gps.c
We noticed that when using the ISR in the intel_mind_gps some of the
rising edges were missed by the platform witch caused the HSU serial to
get stuck at the D0 state, and the platform was not able to go to S0i3.

The solution is to configure host_wake pin as a "wake_gpio" of hsu_port1
(used for gps)in platform_hsu.c.

Change-Id: I623bdaee28e9c317ec9fa868add488fa7c169bb9
Tracked-On: https://jira01.devtools.intel.com/browse/AW-44
Signed-off-by: thamdix <tahax.hamdi@intel.com>
Reviewed-on: https://android.intel.com/501658
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Tested-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>