intel-mid-reboot: avoid sleeping in atomic context
While in the late stage of the reboot process, we are
in atomic context with only CPU0 running.
We have today two functions related to rpmsg that can
sleep: intel_scu_ipc_update_register and
rpmsg_send_generic_simple_command

I'm solving the issue at hand by adding two new variants
of the functions that are "atomic" and unsafe for general usage.
The new variants are missing some needed locks... but since
only CPU0 is online and we are running in atomic context,
we should be pretty safe.

Change-Id: Ieb6e6492430e8673b4c521aedd664e73ed73abc0
Tracked-On: https://jira01.devtools.intel.com/browse/AW-568
Signed-off-by: Guillaume Ranquet <guillaume.ranquet@intel.com>
Reviewed-on: https://android.intel.com/503791
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>
Reviewed-by: Akue, LoicX <loicx.akue@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>