hsu: propagate CTS change to upper layers
When the hsu is configured to be woken up on external GPIOs as in the case
of OHRM and GPS, we can end up in a situation where the tty layer is not
aware of the changes in CTS value when resuming from suspend.

This can lead to a lock of the hsu communication as the tty layer will not
proceed with the TX as long as it thinks the CTS is de-asserted.

To fix the issue, Do propagate the CTS changes to the tty layer whenever
check_modem_status() is called.

Change-Id: If7a78734fd66410fcfc378c0f420d20d66b3b292
Tracked-On: https://jira01.devtools.intel.com/browse/AW-880
Signed-off-by: Guillaume Ranquet <guillaume.ranquet@intel.com>
Reviewed-on: https://android.intel.com/511654
Reviewed-by: Rebraca, DejanX <dejanx.rebraca@intel.com>
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Nassiet, GaelleX <gaellex.nassiet@intel.com>
Reviewed-by: Korpershoek, MattijsX <mattijsx.korpershoek@intel.com>
Reviewed-by: Akue, LoicX <loicx.akue@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>