i2c-designware: add module parameter for SDA hold time
The parameter hold_time is introduced to set the hold time of SDA during
i2c data transfer.
This parameter is read only and can be set in the command line.
The default value is 0x05 according to the soc reference manual.

Example:
add in the command line i2c_designware_core.hold_time=40
(1 / 166 MHz) * 40 = 240 ns of hold time

Change-Id: I45ca0cfc632cf4fdde4624700e56ff2820d5d582
Tracked-On: https://jira01.devtools.intel.com/browse/AW-921
Signed-off-by: Julien Masson <julienx.masson@intel.com>
Reviewed-on: https://android.intel.com/522595
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>
Reviewed-by: Slimani, Kamel <kamel.slimani@intel.com>