display: debugfs: initialize dbgfs_dsi_config auo4x4
The debugfs of the auo4x4 display driver allows reading and writing in
the display controller.
To do so, we rely on a copy of the dsi_config structure.
This copy, named dbgfs_dsi_config, was initialized in the reset()
function.
Since we don't call reset() on boot anymore, dbgfs_dsi_config is always
set to NULL, resulting into the debugfs being useless.

This patch moves the initialization of dbgfs_dsi_config to the
auo4x4_cmd_controller_init() function, which seems more appropriate.

Change-Id: I4b8d9e8b1042288e715a35cc07ced3cd84dff88f
Tracked-On: https://jira01.devtools.intel.com/browse/AW-951
Signed-off-by: Julien Masson <julienx.masson@intel.com>
Reviewed-on: https://android.intel.com/518033
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Tested-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Gong, Sophia <sophia.gong@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>