lsm6ds3h: add probe retry mechanism
Sometimes after a cold boot, the Who-Am-I read register failed. Because
the device seems to be in "sleep mode" and need to receive any commands
to wake up.

This patch changes probe return code to -EPROBE_DEFER to inform the
kernel that this driver requests to be probed on a later time.

Change-Id: I3bdc6c05a5c66ec4c12a81836ba69b67ed776567
Tracked-On: https://jira01.devtools.intel.com/browse/AW-949
Signed-off-by: Julien Masson <julienx.masson@intel.com>
Reviewed-on: https://android.intel.com/518035
Reviewed-by: Hu, Bingquan <bingquan.hu@intel.com>
Reviewed-by: Qin, Jian Feng <jian.feng.qin@intel.com>
Tested-by: Hu, Bingquan <bingquan.hu@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>