gps: drive GPS_MCU_REQ in suspend/resume handlers
mcu_req(on TNG side) is the pin that controls sleep mode for
gps:mcu_req.

The protocol between the SoC and the GPS chip requires this gpio be be
set high when the serial port is active, ie. resumed, and driven low
when the serial port is shut down, ie. suspended.

Change-Id: Ic2030c408300ff7fee0573b92ebab55a8a1f0441
Tracked-On: https://jira01.devtools.intel.com/browse/AW-1003
Signed-off-by: Robert Jarzmik <robert.jarzmik@intel.com>
Reviewed-on: https://android.intel.com/519353
Reviewed-by: Crippa, Roberto <roberto.crippa@intel.com>
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Hamdi, TahaX <tahax.hamdi@intel.com>
Tested-by: Hamdi, TahaX <tahax.hamdi@intel.com>
Reviewed-by: Ranquet, Guillaume <guillaume.ranquet@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>