display: power on panel from deep standby back
After DC enters into SUSPEND mode and then OFF mode in order, it can
not power on panel again, which causes panel being black until reboot.
When setting power mode on and the panel is off, execute the power on
sequence of the panel instead of just exiting low power mode.
In the same way, if setting the power mode to low but the screen is
already off, no need to execute the low power sequence.

Change-Id: Ib32cb1a1dc77ed27dd5a4307416efe43517da05a
Tracked-On: https://jira01.devtools.intel.com/browse/AW-1106
Signed-off-by: wenshelx <wenshengx.wang@intel.com>
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>
Reviewed-on: https://android.intel.com/522735
Reviewed-by: Sun, TianhuaX S <tianhuax.s.sun@intel.com>
Reviewed-by: Gong, Sophia <sophia.gong@intel.com>
Tested-by: Chu, XiaoxiaX <xiaoxiax.chu@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>