bq25898_charger: optimize ADC conversion
The driver is currently waiting for ADC conversion by reading the
status bit and sleeping 100ms, and with a retry loop (NR_RETRY_CNT
being set to 3). Since the first read is done right after triggering
the conversion, it usually show the conversion is not over and so the
driver will sleep at least 100ms.
The experiment show the conversion takes around 80ms so it would be
better to sleep only 30ms and try to read more often. It will let a
chance to the driver to sleep less and the global time is kept to
300ms to avoid any timeout.

Change-Id: Ib347ff6dee9801cadb3966c98ab13982712714c7
Tracked-On: https://jira01.devtools.intel.com/browse/AW-896
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>
Reviewed-on: https://android.intel.com/525195
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Chaumette, HubertX <hubertx.chaumette@intel.com>
Reviewed-by: Ghaddab, RiadhX <riadhx.ghaddab@intel.com>
Reviewed-by: Ranquet, Guillaume <guillaume.ranquet@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>