pwm: intel-mid-pwm: deals with gpios for pwm
When disabling pwm via sysfs, we observed that the pin stays
low or high, depending on the value before disabling.
This behaviour is not expected.

This patch hacks into the pwm driver to temporary switch back to
GPIO mode when disabling pwm indexed by hwpwm.

The gpio number is platform specific. So this patch uses the
get_gpio_by_name to determine it.

Change-Id: I2a87657479c8cb95a80d48a475e7bf171f2e71ed
Tracked-On: https://jira01.devtools.intel.com/browse/AW-1709
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>
Signed-off-by: Afonso PhilippeX <philippex.afonso@intel.com>
Reviewed-on: https://android.intel.com/567651
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Deverge, Jean-francoisX <jean-francoisx.deverge@intel.com>