sensor: fix the gyroscope reports wrong value bug
The gyroscope reports 20rad/s on Z axis when device is idle
and it should be 0. This is very easy to reproduce when tilt
or step counter motion is happend.

The issue comes from the interrupt handler, which read
the fifo data even when the irq is not triggered by fifo
"need_to_read" flag, such as tilt motion happened, this will
make the fifo data become corrupted.

The patch fixes the bug by only reading the fifo when
triggered by fifo "need_to_read" flag.

Change-Id: Iccb03e4ab8097b91cd7b102fe4d58a6ff0f3c40f
Tracked-On: https://jira01.devtools.intel.com/browse/AW-1413
Signed-off-by: Peiqian Li <peiqianx.li@intel.com>
Reviewed-on: https://android.intel.com/535121
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Chen, Shawn Q <shawn.q.chen@intel.com>
Reviewed-by: Hu, Bingquan <bingquan.hu@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>