[WORKAROUND] Remove BESL support in USB
When enable the BESL, the USB configuration is failing quite often at
the SET_CONFIGURATION step. It appears the USB controller does not end
the command transfer it received and so the communication between host
and device is hung. The problem is not systematic and is linked to a
specific timing.
By reverting the support of BESL the USB configuration never fails.

Change-Id: I0b00614c56f50137c86fb15a31d6ed209532e9f6
Tracked-On: https://jira01.devtools.intel.com/browse/AW-1559
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>
Reviewed-on: https://android.intel.com/536235
Reviewed-by: Ranquet, Guillaume <guillaume.ranquet@intel.com>
Reviewed-by: Louis, FabienX <fabienx.louis@intel.com>
Reviewed-by: Akue, LoicX <loicx.akue@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>
Reviewed-by: Ledentec, AlexandreX <alexandrex.ledentec@intel.com>
Tested-by: Akue, LoicX <loicx.akue@intel.com>