lsm6ds3: do not increment fifo_data ptr
fifo_data is incremented in read_fifo() to discard the beginning
of the fifo when an overrun has occured and the pattern has been
broken.

fifo_data has been allocated using kmalloc() earlier during probe.
Thus, incrementing this pointer is wrong as it will be lost forever
and we won't be able to call kfree()
Then, we have a significant risk of buffer overflow as we are
incrementing fifo_data at each overrun.

Change-Id: Ie2fc10f059c51392ae423123252bb9046501b433
Tracked-On: https://jira01.devtools.intel.com/browse/AW-1675
Signed-off-by: Guillaume Ranquet <guillaume.ranquet@intel.com>
Reviewed-on: https://android.intel.com/536623
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>