lsm6ds3h: Do not increment fifo_data ptr
fifo_data is incremented in read_fifo() to discard the beginning
of the fifo when an overrun has occured and the pattern has been
broken.

fifo_data has been allocated using kmalloc() earlier during probe.
Thus, incrementing this pointer is wrong as it will be lost forever
and we won't be able to call kfree()
Then, we have a significant risk of buffer overflow as we are
incrementing fifo_data at each overrun.

Change-Id: If84514e7b378d7d1b7c5a10eb1b834cca2e446f7
Tracked-On: https://jira01.devtools.intel.com/browse/AW-1675
Signed-off-by: Guillaume Ranquet <guillaume.ranquet@intel.com>
Reviewed-on: https://android.intel.com/536622
Reviewed-by: Hu, Bingquan <bingquan.hu@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>
Reviewed-by: Ledentec, AlexandreX <alexandrex.ledentec@intel.com>
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Nassiet, GaelleX <gaellex.nassiet@intel.com>
Reviewed-by: Korpershoek, MattijsX <mattijsx.korpershoek@intel.com>