hsu: propagate CTS change only if CTS not controlled by hw
When the hsu is configured to have the CTS controlled by hardware, as it 
is the case for BT port, we have to prevent any  call to 
uart_handle_cts_change() - software propagation of the CTS - , otherwise we 
can face undefined behavior of hsu and blocking of the Tx link.

To prevent such blocking and fix the issue, this workaround adds a check to the 
hw_ctrl_cts flag before calling or not the uart_handle_cts_change() function.

Change-Id: Id466f4f60e864996f80f04a2ba6f714ccf93fae1
Tracked-On: https://jira01.devtools.intel.com/browse/AW-1613
Signed-off-by: Jean Ferraton <jean.regisx.ferraton@intel.com>
Reviewed-on: https://android.intel.com/538139
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Crippa, Roberto <roberto.crippa@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>