lsm6ds3h: disable i2c master support
This patch is copyed from the following patch(the commit ID listed
below), but not 100% of it, the missing changes are:
1. delete file "arch/x86/configs/x86_64_mrfl_defconfig"

commit ID for reference: 53d179189159d9f2d0cc3d4c1a9357982a22d454

Note: They have to be compiled & get in together.

i2c master support is not needed because we do not have any external sensor.
This patch disables the option in the defconfig and makes some reworks
to be able to compile.

Change-Id: Id813eee0435041dbf02476218676521157080d6d
Tracked-On: https://jira01.devtools.intel.com/browse/AW-1794
Signed-off-by: Simon Dubray <simonx.dubray@intel.com>
Signed-off-by: Kai Qiang <kaix.qiang@intel.com>
Reviewed-on: https://android.intel.com/538510
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Hu, Bingquan <bingquan.hu@intel.com>
Reviewed-by: Ledentec, AlexandreX <alexandrex.ledentec@intel.com>
Tested-by: Ledentec, AlexandreX <alexandrex.ledentec@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>