sensor: fix batching issue for ctsverifier test
When set the batch parameter, if the sensor is not activated,
the parameter "watermark" will just be cached and should be
written to HW when sensor is activated.
But in the existing workflow in lsm6ds3h_set_watermark, it
requires the sensor has been activated before set the watermark register.

Change-Id: Ic24f0dbbac07f5d0d76d9ca7bbc60efadf182f42
Tracked-On: https://jira01.devtools.intel.com/browse/AW-1580
Signed-off-by: Liu, YangX D <yangx.d.liu@intel.com>
Reviewed-on: https://android.intel.com/538745
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Tao, JianX <jianx.tao@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>