Power: leave DPA unit enabled
Don't disable the main display pipe of the display controller,
but rather leave it in the state left by IAFW.

On platforms having a display panel, this enables the display
controller to probe the device. Moreover shutting down the
display controller might trigger a glitch on the display, which
is unsuitable if the display has to remain between IAFW logo and
the kernel boot.

Change-Id: I2bce84e4d811a516b9d789b630abfb0ea3fa18e6
Tracked-On: https://jira01.devtools.intel.com/browse/AW-2281
Signed-off-by: Robert Jarzmik <robert.jarzmik@intel.com>
Reviewed-on: https://android.intel.com/545285
Tested-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>