touchscreen: adapt drivers for synaptics on k4.4
This is the last patch for touchscreen dirver for k4.4,
They have to be compiled & get in together.

Make some adaption to run on k4.4

The patch mainly change an interface introduced by the kernel 4.4:
1. the attribute settings are more strict:
__ATTR(_name,_mode,_show,_store): the "_mode" value don't allow set the
"S_IWOTH" bit. so "S_IWUGO" can't be used anymore, replace it with "S_IWUSR".

Change-Id: I167b2fec15b0605032d0da0a970eb61d44863b8b
Tracked-On: https://jira01.devtools.intel.com/browse/AW-2000
Signed-off-by: Peiqian Li <peiqianx.li@intel.com>