ASoC: Intel: atom: pci: allocate memory for shim regs
The sst's suspend() function saves the shim registers in order
to restore them when the device resumes. The registers are also saved
once, when calling sst_configure_runtime_pm().

These registers are stored in intel_sst_drv->shim_regs64 struct which
is not initialized when we probe the sst in pci mode.
This means that intel_sst_probe() will panic.

For acpi, the memory allocation is done, see sst_acpi_probe()

This patch allocates memory for the shim_regs64 in the probe function.

Change-Id: Ifae53cdcc0dbe53fa163840c217abf049a5c5ac2
Tracked-On: https://jira01.devtools.intel.com/browse/AW-2270
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>
Reviewed-on: https://android.intel.com/544576
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Wei Yu Neng, Loic <loic.wei.yu.neng@intel.com>
Tested-by: Wei Yu Neng, Loic <loic.wei.yu.neng@intel.com>