ASoC: Intel: atom: move context_init() after pci_enable()
sst_context_init() initializes the sst's IRQ interrupt line and
IRQ handling.
However, the associated PCI device is not enabled, which means
that the hardware might not be initialized yet.

This patch moves sst_the context_init() after the pci_*() calls
in order to be able to register the IRQ line.

Change-Id: Ib21ff0045f96634600a0e3b9d2a7e1702e0f6176
Tracked-On: https://jira01.devtools.intel.com/browse/AW-2270
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>
Reviewed-on: https://android.intel.com/544577
Reviewed-by: Jarzmik, Robert <robert.jarzmik@intel.com>
Reviewed-by: Guiriec, Sebastien <sebastien.guiriec@intel.com>
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Deverge, Jean-francoisX <jean-francoisx.deverge@intel.com>
Reviewed-by: Wei Yu Neng, Loic <loic.wei.yu.neng@intel.com>