em: implement wakelock in bq25898 driver
When connected to DCP, bq25898 charger is suspended and watchdog timeout
expires which reinitializes default values.
Charge is stopped.
By implementing a wakelock in bq25898 driver, charger does not suspend
anymore while plugged in DCP and charges to 100%
Wakelock is locked only on DCP connection, nothing changes on SDP
connection.

Change-Id: Ia9a104e799f4482ecebc8a41b085a8d89df1b6d2
Tracked-On: https://jira01.devtools.intel.com/browse/AW-1658
Signed-off-by: MorganX Binet <morganx.binet@intel.com>
Reviewed-on: https://android.intel.com/544824
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Ghaddab, RiadhX <riadhx.ghaddab@intel.com>
Reviewed-by: Chaumette, HubertX <hubertx.chaumette@intel.com>
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>