dwc3-host-intel: fix suspend flow and cleanup
dwc3-host.2 systematically fails to suspend when hcd is in HC_STATE_HALT :
[  758.664998] PM: Device dwc3-host.2 failed to suspend: error -22
And suspend/resume end in a fabric error because we access registers
of the powered down root hub, so do nothing in that case.

At the same time, remove dwc3-host-intel.c inclusion in xhci-plat.c and
implement init and exit functions in the driver.

Change-Id: I35e4ab7d7bde593065a9d10bb99d75c3d8bf68ee
Tracked-On: https://jira01.devtools.intel.com/browse/AW-2212
Signed-off-by: Simon Dubray <simonx.dubray@intel.com>
Reviewed-on: https://android.intel.com/545191
Reviewed-by: Lachaud, EtienneX <etiennex.lachaud@intel.com>
Reviewed-by: Korpershoek, MattijsX <mattijsx.korpershoek@intel.com>
Reviewed-by: Hadjimegrian, MathieuX <mathieux.hadjimegrian@intel.com>
Reviewed-by: Wei Yu Neng, Loic <loic.wei.yu.neng@intel.com>