ASoC: Intel: cdp: trigger ASRC in dmic fixup
On CDP boards we have a clock mismatch between the DMIC and the on
SoC's DSP.
The DMIC expects a clock of 16kHz but the DSP provides 16,667kHz.
This is a SoC limitation, which can only derive of 19,2MHz for devices
plugged on SSP{0,1,2}.

Since the DMIC *must* have 64 bit clocks in each stereo frame, we have
the following clocking configuration:
- base clock: 19.2MHz
- bit clock cycles per frame: 64
- frame sync frequency: {8,16,48} kHz
  -> 19,2 MHz / (SCR*64) where SCR must be an integer and the result
  must be close to 8,16,48} kHz. We have chosen SCR=17 so:
  FS = 19,2MHz / ((17+1) * 64) ~= 16,666666666666668 kHz.

Since we should sample at 16KHz, we have clock drifting.
This is compensated within the DSP firmware by an ASRC.
To trigger the specific mode, we have to request a
frame_size of 64 to the DSP when in i2s mode.

This patch implements the request of frame_sync_width = 64 from
within the machine driver.

Notes:
- this is based on previous work done here:
  Ib20a1fbf0d23c7bea8164cae1bdee12199e69325
- this requires a compatible FW version (01.07.80.00+)

Change-Id: I676d727ea9013c268f4df1fd0322d60cfa97e18c
Tracked-On: https://jira01.devtools.intel.com/browse/AW-2012
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>
Reviewed-on: https://android.intel.com/544955
Reviewed-by: Wei Yu Neng, Loic <loic.wei.yu.neng@intel.com>