sensor: fix the CTS SensorTimeStamps fail issue
The 'lsm6ds3h_enable_accel'function have bug when dealing with
some special cases, and here is a typical example:
    1. Accel and accel_wk are both enabled in sequence
    2. after a while, disable the accel_wk
    3. and then disable the accel
The right logic in step 2 is: clear the accel_wk_on flag,
and keep the odr(because they share a common hardware).
But the current code set the flag based on whether the odr is zero,
and obviously the odr is not zero here, so it lead to this issue.

The patch change it by setting the flag only based on
the 'enable' parameter deliverd by 'lsm6ds3h_enable_accel'function.

Change-Id: I6bda9793f6d0017778fb7435c3c2c854454a0736
Tracked-On: https://jira01.devtools.intel.com/browse/AW-1578
Signed-off-by: Peiqian Li <peiqianx.li@intel.com>
Reviewed-on: https://android.intel.com/545292
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Nassiet, GaelleX <gaellex.nassiet@intel.com>
Reviewed-by: Hu, Bingquan <bingquan.hu@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>
Tested-by: Nassiet, GaelleX <gaellex.nassiet@intel.com>