sensor: adjust st_lsm6ds3h tilt algo loading sequence
fix the HAL writes the new config been mistakenly restored to default settings
by sensor firmware.

f/w first run for tilt will do a default restore (this is the
behavior for f/w), when we start system and HAL has requested the new value
update through driver, at this time f/w for tilt is not actually run once
(we don't have a real tilt triggered), so the new configure is written
successfully, when we do the real tilt actions next, the value we set is
mistakenly restored because the tilt is FIRST time run. So the workaround
in driver is to let tilt enable&disable early. Then next time tilt runs
it will not cause restore.

Change-Id: Id1b0f5112a802274c695390f4c201f3e8766f345
Tracked-On: https://jira01.devtools.intel.com/browse/AW-2011
Signed-off-by: Liu, YangX D <yangx.d.liu@intel.com>
Reviewed-on: https://android.intel.com/545510
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Tao, JianX <jianx.tao@intel.com>
Reviewed-by: Qin, Jian Feng <jian.feng.qin@intel.com>
Reviewed-by: Hu, Bingquan <bingquan.hu@intel.com>
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Tested-by: Jabbari, OussamaX <oussamax.jabbari@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>