display: decrease DSR_COUNT in order to sleep more
DSR count had been increased from 2 to 15 three years ago in order to make
it less power-agressive and avoid shutting down the display for nothing.
On our platforms, we need to reduce it in order to increase the s0ix residency
in use cases with a few fps. It will allow to sleep more between 2 frames.
As free_count is incremented every 16ms, 2 is maybe a bit too agressive (32ms)
but 4 looks like a good trade-off and should be safe enough.

Change-Id: Ie2e142fb7e25897d144c352e6eb3f2f7d0dd8573
Tracked-On: https://jira01.devtools.intel.com/browse/AW-3042
Signed-off-by: Simon Dubray <simonx.dubray@intel.com>
Reviewed-on: https://android.intel.com/565718
Reviewed-by: Lachaud, EtienneX <etiennex.lachaud@intel.com>
Reviewed-by: Bel Aj Ali, HabibX <habibx.bel.aj.ali@intel.com>
Tested-by: Bel Aj Ali, HabibX <habibx.bel.aj.ali@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>