[REVERTME]ASoC: intel: allocate memory for fw in probe()
Audio LPE firmware is loaded everytime a use case is activated.
A corresponding memory area of 204 kB is allocated during the use-case,
and freed when no use-case is active.

When having huge memory pressure, the kzalloc() might fail.
This blocks the fw loading and thus the use-case.

This patch allocates the 204 kB during probe() to ensure that the
firmware can be loaded once we start a use-case.

Notes:
    - Since the FIRMWARE_SIZE is hard-coded, it *must* be updated
      each time we patch the firmware.
    - This is a workaround, because the whole atom driver code does not
      handle the failing kzalloc() properly.

Change-Id: Iee5c1dd7000c2745b7aad4092b1c0218990a5be8
Tracked-On: https://jira01.devtools.intel.com/browse/AW-2963
Signed-off-by: Afonso PhilippeX <philippex.afonso@intel.com>
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>
Reviewed-on: https://android.intel.com/555503
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Bel Aj Ali, HabibX <habibx.bel.aj.ali@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>