net: wireless: bcmdhd: Fix up the BRCM wifi DHD code
Security Vulnerability fix for memory overflow wifi driver
function wl_cfgvendor_rtt_set_config. In the current fix added
check to validate if the target_cnt is valid or not if it is not valid
then parse error. Since target_cnt can be controlled by user netlink
input which needs to validated at the DHD level.

Bug: 32219255
Signed-off-by: Sudhir Kohalli <sudhir.kohalli@broadcom.com>
Tracked-On: https://jira01.devtools.intel.com/browse/AW-3683
Change-Id: I5cf771c60a6ae8019e5e36571197e2849c572b40
Reviewed-on: https://android.intel.com/560338
Reviewed-by: Louis, FabienX <fabienx.louis@intel.com>
Tested-by: Louis, FabienX <fabienx.louis@intel.com>
Reviewed-by: Korpershoek, MattijsX <mattijsx.korpershoek@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>
Reviewed-on: https://android.intel.com/561355