CHROMIUM: drm/tegra: Fix host1x syncpt array OOB access
We use syncpt id to get the syncpt infos from host1x syncpt
array. And the id can be specified by userspace, so in kernel
we need to make sure the ID from userspace doesn't beyond
the boundary of syncpt array.

Unfortunately our host1x driver doesn't do the check right so
this patch fixes it.

BUG=chrome-os-partner:59595
TEST=System boots OK

Change-Id: Ib9b4ee73d9468dbf6b20932c0b1c77dddd743d10
Tracked-On: https://jira01.devtools.intel.com/browse/AW-4089
Signed-off-by: Mark Zhang <markz@nvidia.com>
Reviewed-on: https://android.intel.com/560885
Reviewed-by: Louis, FabienX <fabienx.louis@intel.com>
Tested-by: Louis, FabienX <fabienx.louis@intel.com>
Reviewed-by: Deverge, Jean-francoisX <jean-francoisx.deverge@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>
Reviewed-on: https://android.intel.com/567955
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Bel Aj Ali, HabibX <habibx.bel.aj.ali@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>