pwm: pwm-intel-mid: remove the pwmchip if no gpios
On C2 devices (with drv2605) the pwm driver is not
able to get gpios because they are not present in the sfi table.

This results to a kernel panic on PWM driver:

BUG: unable to handle kernel paging request at 00008130
[ 10.351028] IP: [<c2362743>] pwm_request_from_chip+0x43/0x80
[ 10.351028] IP: [<c2362743>] pwm_request_from_chip+0x43/0x80
[ 10.351109] *pde = 00000000
[ 10.351192] Oops: 0000 1 PREEMPT SMP
[ 10.351364] CPU: 1 PID: 1 Comm: init Tainted: G W 4.4.20+ #1
[ 10.351397] Hardware name: Intel Corporation Merrifield/SPECTRALITE, BIOS 602 2017.02.13:18.06.12
[ 10.351439] task: dea1c000 ti: dea20000 task.ti: dea20000
[ 10.351481] EIP: 0060:[<c2362743>] EFLAGS: 00010246 CPU: 1
[ 10.351524] EIP is at pwm_request_from_chip+0x43/0x80
[ 10.351562] EAX: 00008124 EBX: de06dcf4 ECX: 00000000 EDX: 00000009
[ 10.351602] ESI: c2bdd1a8 EDI: 00000001 EBP: dea21ed0 ESP: dea21ec4
[ 10.351644] DS: 007b ES: 007b FS: 00d8 GS: 0033 SS: 0068
[ 10.351684] CR0: 80050033 CR2: 00008130 CR3: 1c9a7000 CR4: 001006d0

Indeed the probe function is failing and returning error BUT the
pwmchip is not removed. So pwm sysfs remains exported while the
driver is not probed.
So there is a missing pwmchip_remove(&pwm->chip); at the end of
intel_mid_pwm_probe function.

Plus this patch set drvdata to NULL in case of failure.

And another change is to call pci_release_regions function after
that pci_request_regions is done if a failure appears.

Change-Id: Iadb68f9d258f18cd415d9af5d94d4b072e23edee
Tracked-On: https://jira01.devtools.intel.com/browse/AW-4582
Signed-off-by: Afonso PhilippeX <philippex.afonso@intel.com>
Reviewed-on: https://android.intel.com/572195
Reviewed-by: Korpershoek, MattijsX <mattijsx.korpershoek@intel.com>
Reviewed-by: Samoun, JacquesX <jacquesx.samoun@intel.com>
Reviewed-by: Bel Aj Ali, HabibX <habibx.bel.aj.ali@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>